Have user provide minimum possible format version
This additional parameter allows an implementation of repair, without
having to figure out whether omitt_unsplit_suffix needs to be set,
or whether the version information needs to be moved. Doing so is
not impossible, but it is better to provide some repair functionality
for users that never created stores that old.