Clear the range inside the updateStoreHeaderAsync block
updateStoreHeaderAsync uses beginRecordStoreStateWrite to protect
against concurrent updates to the state. Doing it in the `thenRun`
meant that it would:
1. do the clear when going from DISABLED to DISABLED
2. If a race condition occurred, it could do the clear after the
state had been updated. Since there is no way to go from DISABLED
to anything else, I don't think there can be an actual behavior
issue with said raice, but it still seems not great.