resolves #3344: multi phase planning (#3328)
This PR introduces multiphase planning in the `CascadesPlanner`. A
planner phase transitions between planner stages:
```
stages: INITIAL ---------------> CANONICAL ---------------> PLANNED
phases:            REWRITING                   PLANNING
```
                 
The current planner's work is now realized as a planner phase called
`PLANNING` that transitions from planner stage `CANONICAL` to planner
stage `PLANNED`.
`Reference`s maintain the planner stages of the phase that created them
and can be modified by calling `Reference.advancePlannerStage()` which
is currently done during `ExploreGroup`. `CascadesPlanner.Task`s use a
planner phase to resolve the rule set and the cost model that belong to
the phase.

This PR also introduces the `REWRITING` phase which is to be filled out
in subsequent PRs. Currently the ruleset of that phase only contains
`FinalizeExpressionsRule` which is required to make progress and to
allow us to transition to the next planner stage and in turn next
planner stage.

When any phase is initiated in the planner, we will explore the
expression DAG in much the same way as before using the corresponding
planner rule set whose rules then yield new expressions. A rule is
either an `ExplorationRule` or an `ImplementationRule`. When an
`ImplementationRule` yields an expression it is considered a final
expression and is treated exactly the same way as
previously`RecordQueryPlan`s. That means we optimize that final
expressions' inputs using the planner phase's cost model. For
`REWRITING` we use `RewritingCostModel` which consistently picks the DAG
with smallest semantic hash. In that way we construct a DAG bottom up
consisting of the cheapest expression DAG according to
`RewritingCostModel`. The first task to be pushed and the last task in
the task stack will eventually initiate `PlannerPhase.PLANNING` which
then kicks off another round of regular planning.

There are other changes in this PR as well that are necessary in the
same unit of work:
- `References` distinguish between _exploratory_ and _final_ members.
Exploration rules yield and memoize exploratory expressions while
implementation rules yield and memoize final expressions.
- Final expression members are treated in the same way as previously
`RecordQueryPlan`s. An exploratory expression cannot be a
`RecordQueryPlan`
- `Memoizer` has been refactored to be more safe and hopefully easier to
be used. There were a few problems in the code not calling the right
methods for memoization that were fixed.
- All phases explore and implement `RelationalExpression` except the
last `PLANNING` phase which produces `RecordQueryPlan`s. There is a
non-trivial amount of code that makes that somewhat less visible and to
some degree polymorphic.
- most rules now extend `ExplorationCascadesRule` or
`ImplementationCascadesRule` which also determines which yields and
memoization API are available to the rule. Did I mention that I wanted
to make this more safe?
- During `OptimizeGroup.execute()` when we prune the group we previously
would just pick a `RelationalExpression` as the sole survivor if there
were no plans. That logic is changed to only affect final expressions,
i.e. exploratory expressions are kept in the group even after pruning.
If there are no final expressions in the reference before pruning, there
will not be a final expression in the reference after pruning.
- improved Javadocs