Add RecordCountKey State (#3334)
This adds a mechanism for controlling the state of the _deprecated_
RecordCountKey, that can be set on the RecordMetaData.

Note: This is different from `COUNT` indexes.

Resolves #3326 

As noted in the javadocs:

The new state is controlled by
`FDBRecordStore.updateRecordCountStateAsync`, and allows:
1. Changing a READABLE RecordCount to WRITE_ONLY
2. Changing a WRITE_ONLY RecordCount back to READABLE
3. Changing any state to DISABLED

The default state is READABLE, and it will stay READABLE unless
explicitly updated.
If the RecordCount is WRITE_ONLY then it will maintain the count, but
not allow querying it.
If the RecordCount is DISABLED it will clear the data, and stop
maintaining the values.


In order to support this new state, a new `FormatVersion` has been
added: `RECORD_COUNT_STATE`
In order to use the new functionality, the stores must be built with at
least this format version.

---------

Co-authored-by: Alec Grieser <alloc@apple.com>