New planner rule added to push predicates down in the query representation (#3324)
This introduces a new query planner rule that will be a part of a larger
framework to canonicalize query representations. The rule pushes down
predicates to ensure that they appear as low in the tree as possible.
So, for example, suppose you had a query like:

```sql
SELECT a, b, c
   FROM (SELECT a, b, c, d FROM T)
   WHERE d = ?
   ORDER BY a, b
```

This will produce a new query expression where the `d = ?` is pushed
lower in the graph, something like:

```sql
SELECT a, b, c
   FROM (SELECT a, b, c, d FROM T where d = ?)
   ORDER BY a, b
```

This simplification helps us during planning to push down predicate
execution to the lowest part of the graph where they can go.

Note that for the moment, this rule has _not_ been added to the default
`PlanningRuleSet`. The reason for this is that some of the rewrite rules
will have weird overlap with some of the other rules, and so we want to
rewrite the query in one phase and then plan and optimize it in another.
The framework for that is not there yet