WIP: Attempt to integrate the rewrite rules into the planner
This attempts to integrate the rewrite rules into the planner. At present, those are the PredicatePushDownRule and the SelectMergeRule.

While integrating those rules, we saw a significant regression amongst plans that used SQL functions. Particuarly, the sql-functions.yamsql file started timing out, and many of its tests ended up taking several seconds all on its own.

I did some experimentation since then to try and figure out what was going on. There's an explicit cross-product in the SelectMergeRule that seemed a little suspect (in particular, it would try and merge all of its mergeable children, but if it had multiple mergeable children, it would create one new variant for each element in the cross product of the original children). I attempted to limit the number of children by having it instead pick the best of its children, and that helped somewhat, but sql-functions.yamsql in particular was very slow.

I then tried a different tack. One thing about selection merge is that it will turn a tree of joins into a single n-way join. Our join planning will then do a full exploration, and in particular, with SQL functions, we get additional "values" boxes. Those then participate in join planning, but because they are uncorrelated, there are no constraints in the places they can go, which leads to a combinatorial explosion of its own. If I add an additional rule which prunes out a quantifier over `tf(range(1))` (which isn't legal if that's the only quantifier, but should be fine otherwise as long as nothing is correlated to it), then that simplifies the query enough that we can do significantly less join planning. This seems to put us on roughly the same footing, performance wise, especially if we also try to limit the variants of merged queries.