Don't memoize correlation sets in planner objects (#3365)
We should not use a memoizing supplier to cache the correlation sets of
a quantifier since the reference the quantifier ranges over is mutable
with respect to the contained expressions. While these expressions are
rarely exposing new correlations it may be that a new expression is
added that has no correlations at all. If we prune to that expression
later, the correlation set, if memoized, is wrong.

Example
```
SELECT *
FROM A, (SELECT * FROM B WHERE A = 5 AND FALSE)
```

The inner
```
...(SELECT * FROM B WHERE A = 5 AND FALSE)...
```
has `B` in its correlation set. That expression, however, can also be
simplified to:
```
...(SELECT * FROM B WHERE FALSE)...
```
(and even further). The correlation set is empty.

This PR deals with the change of removing the `MemoizingSupplier` as
well as the fallout from now exposed bugs.