Create throttled and retrying record iterator (#3350)
This PR creates the `ThrottledRetryingIterator` for use with cases that
need a reliable retrying mechanism with resource controls.
for the most part, this is new code that is not in use yet. Small
changes to the existing code base (log message fields, executor method
made public, refactoring of the FutureAutoClose) were made as well.
This class is meant to live at the same level as other
`FDBDatabaseRunner`s even though this class does not implement this
interface. Hence, this is using the `TransactionalRunner` and
`FutureAutoClose` classes instead of the (already retrying)
`FDBDatabaseRunnerImpl`.
The iterator implements resource controls - ops/sec and ops/transaction
for scan and delete operations. It does not implement similar controls
for other operations (update, create etc). These can be added later when
they are needed.

Resolves #3355