Allow `CASE` alternative implicit type promotions when feasible. (#3379)
Previously, the `CASE` runtime used to reject alternatives with
different types even if they have a common max-type.

This PR address this by implementing an encapsulator that is capable of
generating a `CASE` expression callsite that with implicit promotions
applied to the alternatives when necessary.

As a result a query like this now works:

```sql
create table A(A1 bigint, A2 bigint, A3 bigint, primary key(A1));
insert into A values
        (1, 10, 10),
        (2, 11, 20),
        (3, 12, 30);
select a3, case when a3 > 15 then 10 else 3.14 end from A;
10, 3.14
20, 10.0
30, 10.0 
```

(note how `10` is promoted to `10.0` in the result set).

This fixes #3378.