Planner now uses correct calculation for ordering of aggregate scans over permuted min/max indexes (#3333)
The way the ordering key is generated for permuted min and max indexes
requires that the planner handle the permutation of the min/max
aggregate with zero or more grouping columns. We were not correctly
handling that, which meant that we were crafting incorrect planner keys.
This would lead to strange comparison keys with IN-union plans, and it
could lead to `ORDER BY` constraints not being matched. This handles
that appropriately, with some additional tests to make sure we are now
considering more interesting examples of queries with min or max indexes
and ordering constraints.

This fixes #3331.