Call Store.Builder.openAsync instead of build (#3375)
Call `openAsync` on the store builder and adjust tests accordingly.
The store initialization sequence was adjusted to the asynchronous
nature of the store creation (best viewed using "hide whitespaces").
The tests were adjusted in two ways:
1. Since the store is now required to exist for the iterator to do its
work (not a burden since it is assumed that the cursor creation would
need a store anyway), a `commit()` was added to the store opening in the
tests. Since most tests create a fake cursor from a list of integers,
this was not needed before.
2. Because the async initialization sequence now has a future for the
store (as opposed to a concrete instance before), the first future of
the iteration is not created until the store future is completed. Some
of the tests assert that the first `handleItem` will be called by the
time the iteration starts, which is not guaranteed anymore. To resolve
this, a delay has been introduced to allow the future to get scheduled.
It may lead to flaky tests in PRBs, and we may need to modify the tests
if that happens.

Resolves #3377