(non)-partitioning properties for `Reference`s (#3382)
This PR provides a distinction among the properties that are being
maintained for `References`. Before this PR, all properties were only
used to derive `ExpressionPartition`s, i.e. they would define a set of
properties all plans in the reference could be partitioned by. There
are, however, also other properties which should be maintained in the
same way but should be accessible individually for each expression in
the reference.