Fix issue with checkVersion getting record store state is being used for queries (#3386)
Resolves #3384 

As described in the issue, modifying the state in this way is not good.
I opted against refactoring as part of this PR, and instead did the
tactical approach. I did create a draft PR that helps validate that this
is the only place that was actually updating the reference:
https://github.com/FoundationDB/fdb-record-layer/pull/3385 but I don't
think that refactoring is worth doing at this point, and suspect there's
something better, but I didn't want to delay this bug fix to come up
with a design.

I found the original test that came with the original change, and added
a comment there too, but didn't find it until after I had a separate
test, I feel the new test covering the `IndexQueryabilityFilter` is
different enough that having both is worthwhile.