Record repair: External Runner API and repair operations (#3370)
This PR adds the API to run the repair operations.
The API has 2 entry points: `runStats` and `runValidateAndRepair`. The
idea is that the first one gives an aggregate count of found issues
whereas the second one returns a (size bound) list of issues.
This PR also adds the ability to run repair operations (in addition to
the previously existing validate operations).
A note on tests:
In order to help with readability, the end-to-end tests for the feature
are divided into `validate only` and `validate and repair`, where the
former test the validation only (all issues are detected) and the latter
adds repair operations. For the repair flow, a subsequent validate and
load of records is generally added to verify the post-repair state of
the store.
In addition, more localized (though not quite unit tests) were modified
to handle the validators logic in isolation.
resolves #3372