Allow version repair to operate while the store is locked (#3396)
Create a version of the `saveRecordAsync` that allows bypassing the new
store update lock.
We could have made this more complicated by creating a framework for
rules around when can a certain kind of lock be bypassed and for what
"reason" (e.g. "repair"), but this seems like an overkill so I opted for
the straightforward way of a single boolean and an `if (override)` in
the base `saveRecordAsync` implementation.

Resolves #3391