Temporary SQL Functions (#3364)
This implementation provides support for temporary SQL functions that
are transaction-bound, meaning their scope is limited to the duration of
the transaction in which they are defined. Similar to regular SQL
functions, temporary SQL functions are pure metadata artifacts,
therefore, they manifest as additional information attached to the
metadata object `RecordLayerSchemaTemplate`.

When a user defines a temporary SQL function, a newly created
(temporary) `RecordLayerInvokedRoutine` is added to the
`RecordLayerSchemaTemplate`, and the resulting metadata object is
attached to the active `FDBRecordContext`’s session map. During
statement parsing, the plan generator first searches for the
`RecordLayerSchemaTemplate` object in the transaction’s session. If it
is not found, it falls back to using the `RecordLayerSchemaTemplate`
object provided in the planning context, effectively implementing the
transaction-bound metadata visibility.

Since transaction-bound metadata is not covered by the schema template
versioning scheme, it necessitates specialized handling in the plan
cache. Currently, the plan cache incorporates the schema template
version in the cache key to distinguish plans generated from distinct
schema templates. However, with transaction-bound metadata, it is
possible to generate plans that share the same schema template version,
which is incorrect. Therefore, we require a method to include the
transaction-bound metadata in the cache key alongside the schema
template version.

To achieve this, the transaction-bound metadata is added in its textual
form to the `QueryCacheKey`. Temporary functions are permitted to be
prepared, and their prepared parameters, along with their literals, are
extracted and attached to their definitions within the metadata. This
approach enables the reuse of these functions (in their expanded form)
along with the queries (referencing them) when feasible, i.e., when the
`QueryPlanConstraint`(s) surrounding the corresponding cached physical
plan are satisfied with the set of literals extracted from a newly
executed statement.

This PR identified a bug that allows defining prepared non-temporary SQL
functions. This bug currently exists because there is no support for
serializing the prepared parameters along with the function itself.

Numerous tests are added to verify the correctness of the caching
infrastructure and the proper expansion of temporary functions.

Example temporary function:

```sql
autoCommit=off -- as per JDBC connection
create table t1(pk bigint, a bigint, primary key(pk));
insert into t1 values (1, 10), (2, 20), (3, 30), (4, 40), (5, 50)

-- define the temporary function
create temporary function sq1(in x bigint) on commit drop function as select * from t1 where a < 40 + x

select * from sq1(x => 2)
1, 10
2, 20
3, 30
4, 40

rollback -- as per JDBC connection

select * from sq1(x => 2) -- ERROR! the function is not visible anymore.
```

This closes #3363.