Introduce a new planner rewrite rule to facilitate in-lining function parameter values (#3381)
This adds the `DecorrelateValuesRule`. It operates on a single select
expression to remove any values boxes (potentially introduced during
function expansion). It will rewrite the immediate select expression's
result value and predicates, and then it will rewrite its children so
that each child now "owns" a copy of the values box(es) instead. By
repeated application of this rule, the expression can be simplified so
that there are no values boxes remaining.

Note: this does not currently take much care to make sure that each
value is "pure". If there were an expression with a side-effect (like,
say, `rand()`) that could not be pushed down, this would make some
illegal transformations. Modifying this rule to exclude those values
should be a relatively straightforward future modification, assuming one
has a good oracle for function referential transparency.

This resolves #3380.