Introduce `SelectMergeRule` (#3388)
This PR introduces `SelectMergeRule` as a new rule that works as an
implementation rule in the `REWRITING` phase of the planner.

The idea behind the `SelectMergeRule` is to merge `SelectExpression`s if
they are adjacent to one another in the query graph, thus reducing
overall complexity of the query graph and allowing proper join
enumeration during the `PLANNING` phase.

---------

Co-authored-by: Alec Grieser <agrieser@apple.com>