`EvaluationContext` and constraints for simplifications (#3387)
This PR is a direct precursor of enabling a `PredicateSimplification`
rule in the `REWRITING` phase of planning.

This PR gives all planner rules including
- cascades planner rules
- simplification rules for `Value`s and `QueryPredicate`s
- computation rules for `Value`s and `QueryPredicate`s

the ability to access the current evaluation context and to produce
`QueryPlanConstraint`s as part of the result of the rule when yielding.

Yielding has been extended to allow to yield `QueryPlanConstraint`s with
the object that is being yielded.

For conciseness and readability I added a fluent-like API for rules to
use:

```
call.yieldResultBuilder()
    .addConstraintsFrom(bindings.get(getMatcher()))
    .addConstraintsFrom(Sets.difference(allMajorOfMinors, retainedMajorOfMinors))
    .yieldResult(simplifiedPredicate);
```

Note that for the propagation to work, we need to make sure that in each
individual rule, objects (i.e. normally in a tree of `QueryPredicate`s
or `Value`s) that are not being preserved by the rule need to have their
constraints preserved.

Example:

`COV(c58) IS NOT NULL AND FV(QOV(q, "x") = Literal(5)`

can be simplified to if `c58` is bound to `42`:

`TRUE AND FV(QOV(q, "x") = Literal(5)` with constraint that `COV(c58) IS
NOT NULL` which is assigned to `TRUE`

that can be further simplified to:

`FV(QOV(q, "x") = Literal(5)` with constraint that `COV(c58) IS NOT
NULL` which now needs to be assigned to the `ValuePredicate` as `TRUE`
is gone. That means that `IdentityRule` must preserve the plan
constraints that was previously assigned to the `TruePredicate`.