Planner rewrite rule for simplifying `SelectExpression` predicates (#3411)
This introduces a new planner rewrite rule called
`QueryPredicateSimplificationRule`, it matches `SelectExpression`s with
at least one predicate, and attempts to simplify the conjunction of the
predicates through the `QueryPredicate` simplification engine and rules
(which were introduced in #3404).

If the predicate was simplified, it proceeds to create a new
`SelectExpression` with that simplified predicate, and if the predicate
is a conjunction, it breaks it apart, and supply its terms to the
`SelectExpression` to align with the way the `SelectExpression` is
constructed where its predicates are heavily preprocessed.

This PR also introduces a new `QueryPredicate` simplification rule
called `ConstantFoldingMultiConstraintPredicateRule` that, as the name
suggests, attempt to simplify a `PredicateValueWithRangeAndConstraints`
that has 1 `RangeConstraints` with _at least_ 2 constraints. There are
some interesting cases where the resulting constraints are contradicting
and result in disjoint singleton ranges that are impossible to satisfy
(more about that explained #3407).

Finally, the PR introduced a new randomized predicate testing framework
that is able to generate complex predicate trees and validate them. It
integrates with the `RandomizedTestUtils` so it can also benefit from
any nightly testing capabilities.

This resolves #3407 and #3410.