Introduce value-specific plan cache constraints (#3413)
This introduces new set of value-specific constraints to be added to the
plan cache keys, the constraints are created for the following:

* Explicit boolean values (true, false).
* Explicit null values.
* Whether a literal is not null.

Implementing these constraints is crucial to prevent the use of physical
plans generated based on assumptions about literal values that may not
hold true during subsequent query executions. This can occur when the
actual literal values differ from those initially assumed.

Therefore, the plan generator should analyze the literals and generate
necessary constraints about their values according to the above rules to
be used in the plan cache key.

This resolves #3412