Add new `Option`s to disable planner rules through the Relational API (#3421)
This exposes the existing methods on the `RecordQueryPlanConfiguration`
to allow us to disable planner rules through the Relational API. It also
adds a new helper method that allows us to turn off all of the planner
rewrite rules at once, which can be useful as a stop-gap as we work to
integrate them further.

This resolves #3420.