Use FileDescriptor to resolve names
This parses the protobuf to find the actual type names, removing
our need to replicate protobufs name resolution.

This does have the downside that it might result in converting
non-fully-qualified types to fully-qualified types when it doesn't
need to, but I'm not going to optimize that at this point.

There's also still a bunch of TODOs because I didn't want to flush
out the whole call hierarchy. And FDBMetaDataStoreTest still fails,
but that will come next.