Scoring and costing query predicates (#3427)
This introduces a simple algorithm for scoring predicate complexity
based on the predicate's tree diameter (defined as the longest path
between any two leaf nodes in the predicate tree, assuming an undirected
tree, as per Cormen et al.'s *Introduction to Algorithms*). To
facilitate this, a `diameter()` method has been added to the `TreeLike`
class for calculating the diameter of a `SelectExpression`'s predicates.

Furthermore, a `PredicateComplexityProperty` is introduced, which
returns the maximum predicate diameter found within a
`RelationalExpression` tree. This property is then leveraged in the
`RewritingCostModel` as a tie-breaker: when expressions have identical
costs, the expression with the smallest maximum predicate diameter is
preferred.

This fixes #3419.