Avoid over-eager Boolean simplification
There is an issue with the Boolean simplification logic in that it was attempting to simplify expressions like:

```sql
WHERE B IS TRUE
```

To:

```sql
WHERE FALSE
```

Where `B` is just some field. I had to do some digging, and this ended up being caused by the `ConstantFoldingMultiConstraintPredicateRule`, which had some logic that was not sufficiently precise when it was considering the possible values and therefore could mistakenly think that two constant types were in the wrong range.

The solution I went with here was to refactor that code so that it used a different method (the same method as was used for single-constraint predicates) which was already correct. I added some tests to that affect in the constant folding test suite. There weren't tests before that would check the simplification of predicates on `UNKNOWN` or `NOT_NULL` values, which would include fields, outside the context of a more sophisticated rules, so not all of the examples I added were failing before the change, but there should now be coverage tof the tests that were failing in `boolean.yamsql`.