Address nullability of null-on-empty quantifiers and nested fields (#3426)
This PR addresses two problems, both of which are necessary to resolve
problems encountered by
`FDBSimpleGraphQueryTest.testSubselectHasNullOnEmptyAndIsNullPredicate`
when trying to integrate the planner rewrite rules (see: #3401).

The first is that `Quantifier.forEachWithNullOnEmpty` quantifiers were
not taking into account the fact that they could return `null` in their
return type. That meant that predicate simplification rules based on
simplifying `IS (NOT) NULL` based on the field type could make incorrect
inferences about the result of such quantifiers. This updates the result
type method for that class so that we now take that into account.

The second is that if a type had non-nullable fields, but it was a
nested type that lived inside a nullable field, then we’d report that
field values on the child field were non-nullable. This is incorrect, as
if the whole struct is `null`, field accesses on the child can return
`null`. This could also result in the same rules making incorrect
inferences about the type. This addresses this by looking at the types
as we walk down field access paths, and then taking into account their
nullability when we return the field value's return type.

I also looked into a different approach, where we’d just pro-actively
mark all fields of a nullable type as nullable. This unfortunately has
cross version compatibility problems with this verify:


https://github.com/FoundationDB/fdb-record-layer/blob/296b4e5b1e1196c92dd78ceda54bc9f82b1dbe4d/fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/plans/RecordQueryDefaultOnEmptyPlan.java#L79

Essentially, if the inner type has some (nested) non-nullable fields,
then the nullable on empty value will differ in more ways than just the
top-level type being nullable. This causes cross-version continuation
tests to fail. We could solve _that_ by loosening the assert to only
require that the types by the same (recursively) ignoring nullability,
and then constructing a new type that combines field nullabilities.

There are also problems with the fact that this erases the nullability
constraints entirely, which means that you wouldn’t be able to assert
that, say, either the whole parent struct is null _or_ all of its
non-null fields are not null.