Separate out Lucene tests so they run separately in PRB (#3440)
I was looking through recent PRB runs, and it seemed like the time to
run other-tests was now around twice the time to run the core-tests.
Most of the time seemed to be coming from lucene-tests, so I wanted to
experiment with separating those out. This hopefully should decrease our
total PRB time