Run expensive `Traversal` verification sanity check less often (#3442)
A sanity check to validate the traversal integrity was recently added in
#3435. That check alone seems to almost double the total time it takes
to run the `core-tests` during PRB from around 15 minutes to around 30
minutes. It's a useful check, but it is probably running too often. We
should be able to get away with only running it occaisionally and then
only enabling it more stringently if the need arises.