Update code version before running tests in release build (#3450)
This modifies the release build workflow so that we increment the
version of local code before we go to run the tests. This makes it so
that we correctly distinguish between the latest external version (which
should have a version less than `!current_version`) and the current code
version. This is necessary to unblock the release build, which is
failing currently due to certain tests failing when run in mixed mode
against the lateset published version due to expected differences.

Overall, I think this is a better set-up for our tests, as the version
that will actually run the tests is now closer to version that we will
actually end up publishing. The only difference should be the release
number, which _shouldn't_ matter, but it's good to be sure.

This fixes #3449.