Pick correct locally generated build for `!current_version` tests
The previous logic for choosing the locally generated build for `!current_version` required that there be the word `SNAPSHOT` in the jar name. This had the effect that during release builds, we wouldn't run tests with the current version as an external version. This resulted in test failures after `ExternalServerTest` was added.

This fixes #3454.