Update the Maven Central publishing URL to avoid sunset service
The service we were using to publish to Maven Central was sunset. There is a legacy API that allows us to continue to use our old publising gradle platform, and so I've updated the configuration so that it follows the advice available on their website: https://github.com/gradle-nexus/publish-plugin?tab=readme-ov-file#publishing-to-maven-central-via-sonatype-central

This is advertised to be pretty straightforward. The two main steps are (1) update the publishing URL to one that uses a bridge API and (2) modify the credentials to ones generated by the new portal. I've done that (and already updated the secrets on the repo with the new credentials). So we should be able to get a new build published through the new gateway once this is in.

This fixes #3456.